requirements.txt
asgiref==3.2.10
autopep8==1.5.4
beautifulsoup4==4.9.1
Django==3.1
django-bootstrap4==2.2.0
django-cachalot==2.3.0
django-comments-dab==1.6.7
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-filter==2.3.0
django-qrcode==0.3
Pillow==7.2.0
pycodestyle==2.6.0
pytz==2020.1
qrcode==6.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
